src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/focus-navigation-with-distributed-nodes-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
This tests that pressing Tab key should traverse into shadow DOM subtrees, and pressing Shift-Tab should reverse the order.

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
